backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.5
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
firebase-admin==6.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.38.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-watch==4.2.0
aiofiles==23.2.1
ffmpeg-python==0.2.0
beautifulsoup4==4.12.2
psutil==5.9.6
supabase==2.20.0
python-dotenv==1.0.0